requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==1.3.5
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.3
openpyxl==3.1.2
email-validator==2.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
